Update V3 client to make it easier to use with api
- The V3 API expects params in 3 different places (path, query, and
body). This allows us to take any params, in a method, but have a
definition in that method that defines what params are allowed where,
and put them in the right place.
- Also, in the V3 API, some params are supposed to be nested inside an
object name param. If we pass an object name, this will handle the
nesting for us.
- See next 2 commits for usage examples
- I need to look back over the V2 docs...this may be something we need
to incorporate there as well.